monitoring: fix scheduler + controller-manager
This commit fixes the discovery of the scheduler and controller-manager
metrics. The underlying issue is that kubeadm does not create Service
resources for these two deployments, meaning that the ServiceMonitor
cannot select and endpoints for Prometheus to scrape. The correct
solution would be actually use kube-prometheus as a jsonnet library and
to generate configuration for the `kubeadm` platform rather than use
the example manifests. As a short term measure, this commit manually
adds the manifests that would have been generated by this configuration,
as defined in [0].

[0]
https://github.com/prometheus-operator/kube-prometheus/blob/main/jsonnet/kube-prometheus/platforms/kubeadm.libsonnet.

Signed-off-by: Lucas Servén Marín <lserven@gmail.com>